Change webpack to a TS file
