+++ .eslintignore
@@ -4,6 +4,6 @@ node_modules
 dist
 .eslintrc.js
 jest.config.js
-webpack.config.js
+webpack.config.ts
 setupTests.ts
 coverage
